// frontend/src/api/openapi/models/TrainingParameters.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TrainingParameters = {
  /**
   * The maximum number of epochs to train the model
   */
  max_epochs: number;
  /**
   * The batch size to use for training
   */
  batch_size: number;
  /**
   * The learning rate to use for training
   */
  learning_rate: number;
};